renamed utils.py -> tokenManager, ValidateUserServie -> validateUserServie and libs -> utils
